requirements/_base.txt
# venv requirements for base env
boto==2.49.0
Django==1.8.19
django-blog-zinnia==0.16  # Django 1.8
django-ckeditor==5.5.0  # Django 1.8
django-colorful==1.2.0
django-contrib-comments==1.8.0  # Django 1.8
django-extensions==1.7.9  # Django 1.8
django-filer==1.3.2
django-haystack==2.7.0  # Django 1.8
django-mptt==0.8.7  # Django 1.8
django-orderable==6.0.0
django-polymorphic==1.3.1  # Django 1.8
django-redis==4.8.0  # Django 1.8
django-robots==3.1.0
django-statsd-mozilla==0.4.0
django-storages==1.6.6
django-tagging==0.4.6
django-widgy[all]==0.8.4
django-xmlrpc==0.1.8
Fabric==1.14.0
gspread==3.0.1
Mezzanine==4.3.1
oauth2client==4.1.2
Pillow==5.2.0
pyOpenSSL==18.0.0
pyquery==1.4.0
python-redis-lock==3.2.0
raven==6.9.0
redis==2.10.6
solid-i18n==1.4.2
Whoosh==2.7.4
